Bug 1580687 - Cleanup new endpoint
